+++ src/main/java/cn/rjys365/sebookstorebackend/dao/daoimpl/UserDAOImpl.java
@@ -16,7 +16,7 @@ public class UserDAOImpl implements UserDAO {
     }
 
     @Override
-    public Optional<User> findUserById(Integer id) {
+    public Optional<User> findUserById(Long id) {
         return this.userRepository.findById(id);
     }
