+++ src/main/java/cn/rjys365/sebookstorebackend/entities/Order.java
@@ -29,7 +29,7 @@ public class Order {
 
     @Getter
     @Column(name="user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name="created_time")
     private LocalDateTime createdTime;
@@ -45,7 +45,7 @@ public class Order {
         this.id = id;
     }
 
-    public void setUserId(Integer userId){
+    public void setUserId(Long userId){
         this.userId=userId;
     }
