+++ src/main/java/cn/rjys365/sebookstorebackend/service/UserService.java
@@ -11,19 +11,19 @@ import java.util.Optional;
 public interface UserService {
     public Optional<UserLoginResponse> loginByUsername(String username, String password);
 
-    public User blockUser(Integer id,Integer blockerId);
+    public User blockUser(Long id,Long blockerId);
 
-    public User unBlockUser(Integer id,Integer unBlockerId);
+    public User unBlockUser(Long id,Long unBlockerId);
 
-    public Iterable<User> getAllUsers(Integer getterId);
+    public Iterable<User> getAllUsers(Long getterId);
 
     public Optional<User> addUser(UserRegisterRequest user);
 
-    public Optional<List<CartItem> > getUserCartItems(Integer userId);
+    public Optional<List<CartItem> > getUserCartItems(Long userId);
 
-    public Optional<List<CartItem> > setUserCartItems(Integer userId, List<CartItem> cart);
+    public Optional<List<CartItem> > setUserCartItems(Long userId, List<CartItem> cart);
 
-    public Optional<List<CartItem> > addUserCartItem(Integer userId, Integer BookId, Integer count);
+    public Optional<List<CartItem> > addUserCartItem(Long userId, Integer BookId, Integer count);
 
-    public Optional<List<CartItem> > setUserCartItem(Integer userId, Integer BookId, Integer count);
+    public Optional<List<CartItem> > setUserCartItem(Long userId, Integer BookId, Integer count);
 }
